[qfcc] Handle signed-unsigned int comparison better
This fixes the upostop-- test by auto-casting implicit constants to
unsigned (and it gives a warning for signed-unsigned comparisons
otherwise). The generated code isn't quite the best, but the fix for
that is next.

Also clean up the resulting mess, though not properly. There are a few
bogus warnings, and the legit ones could do with a review.

+++ tools/qfcc/test/arraytypedef.r
@@ -41,7 +41,7 @@ find_xdef (string varname)
 	//FIXME need a simple way to get at a def's meta-data
 	xdefs_t    *xdefs = PR_FindGlobal (".xdefs");
 	xdef_t     *xdef = xdefs.xdefs;
-	while (xdef - xdefs.xdefs < xdefs.num_xdefs && xdef.addr != varaddr) {
+	while (xdef - xdefs.xdefs < (int) xdefs.num_xdefs && xdef.addr != varaddr) {
 		xdef++;
 	}
 	if (xdef.addr != varaddr) {
